Fix saving new scenes
couldn't fit the full thing in the title, but this fixes the error that would come from saving a file that doesn't already exist in the mod nor the base game files

+++ SonLVLAPI/LevelData.cs
@@ -612,8 +612,9 @@ namespace SonicRetro.SonLVL.API
 		{
 			string fullpath = Path.Combine(ModFolder, "Data/Stages", StageInfo.folder, name);
 			bool isnew = !File.Exists(fullpath);
+			bool noModExists = (DataFile != null && DataFile.FileExists($"Data/Stages/{StageInfo.folder}/{name}")) || File.Exists($"Data/Stages/{StageInfo.folder}/{name}");
 			action(fullpath);
-			if (isnew && ReadFileRawNoMod($"Data/Stages/{StageInfo.folder}/{name}").FastArrayEqual(File.ReadAllBytes(fullpath)))
+			if (isnew && noModExists && ReadFileRawNoMod($"Data/Stages/{StageInfo.folder}/{name}").FastArrayEqual(File.ReadAllBytes(fullpath)))
 				File.Delete(fullpath);
 		}
